[refactor] #22 후보자Util 삭제 후 서비스 클래스로 이동
